requirements.txt
# LangChain Core and Community
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# UI Framework
streamlit>=1.30.0

# Search Tool
tavily-python>=0.3.0

# Environment and Utilities
python-dotenv>=1.0.0
openai>=1.10.0

# Image handling (if using base64 for vision tasks)
Pillow>=10.0.0